+++ packages/eslint-plugin/src/configs/all.ts
@@ -41,6 +41,7 @@ export const all: Linter.BaseConfig = {
           "declaration",
           "defApiFeature",
           "defaultValue",
+          "deferred",
           "detail",
           "end",
           "eventproperty",