+++ packages/mergebot/src/_tests/cachedQueries.d.ts
@@ -1,0 +1,8 @@
+import * as cachedQueries from "../util/cachedQueries";
+
+type CachedQueries = {
+  [key in keyof typeof cachedQueries]: Awaited<ReturnType<(typeof cachedQueries)[key]>>;
+};
+
+declare const _default: CachedQueries;
+export = cachedQueries;